// client/src/Redux/actions/types.tsx
enum ActionTypes {
    LOGIN_START = "LOGIN_START",
    LOGIN_SUCCESS = "LOGIN_SUCCESS",
    LOGIN_FAILURE = "LOGIN_FAILURE",

    CREATE_USER_START = "CREATE_USER_START",
    CREATE_USER_SUCCESS = "CREATE_USER_SUCCESS",
    CREATE_USER_FAILURE = "CREATE_USER_FAILURE",

    START_CHECK_COOKIE = "START_CHECK_COOKIE",
    CHECK_COOKIE_SUCCESS = "CHECK_COOKIE_SUCCESS",
    CHECK_COOKIE_FAILURE = "CHECK_COOKIE_FAILURE",

    SET_LOADING = "SET_LOADING",
    


    LOGOUT = "LOGOUT",
}

export default ActionTypes